refined gRPC library
