+++ src/envoy-proto-decorator.ts
@@ -1,32 +1,35 @@
-import {
-  credentials,
-  ServiceDefinition,
-  Metadata,
-  ServiceError,
-  ChannelCredentials,
-  Client
-} from "grpc";
+import { credentials, ServiceDefinition, Metadata, ServiceError } from "grpc";
 import EnvoyContext from "./envoy-context";
 import EnvoyGrpcRequestParams, { EnvoyGrpcRequestInit } from "./envoy-grpc-request-params";
-import { debug } from "util";
+import {
+  RequestFunc,
+  EnvoyClient,
+  ClientConstructor,
+  EnvoyClientConstructor,
+  EnvoyClientFuncEnabled
+} from "./types";
 
-export interface ClientConstructor {
-  new (address: string, credentials: ChannelCredentials, options?: object): Client;
-}
-
-export type RequestFunc = (request: any, options?: EnvoyGrpcRequestInit) => Promise<any>;
-
-export interface EnvoyClientFuncEnabled {
-  [methodName: string]: RequestFunc | any;
-}
-
-export interface EnvoyClient extends Client, EnvoyClientFuncEnabled {
-  readonly originalAddress: string;
-  readonly envoyContext: EnvoyContext;
-}
-
-export interface EnvoyClientConstructor {
-  new (address: string, envoyContext: EnvoyContext): EnvoyClient;
+function makeAsyncFunc(name: string): RequestFunc {
+  return async function(this: EnvoyClient, request: any, options?: EnvoyGrpcRequestInit) {
+    const params = new EnvoyGrpcRequestParams(this.envoyContext, options);
+    return new Promise((resolve, reject) => {
+      Object.getPrototypeOf(Object.getPrototypeOf(this))[name].call(
+        this,
+        request,
+        params.assembleRequestMeta(),
+        {
+          host: this.originalAddress
+        },
+        (error: ServiceError, response: any) => {
+          if (error) {
+            reject(error);
+            return;
+          }
+          resolve(response);
+        }
+      );
+    });
+  };
 }
 
 /**
@@ -58,31 +61,11 @@ export default function envoyProtoDecorator(
 
   const prototype = clazz.prototype as EnvoyClientFuncEnabled;
 
-  for (const name in Object.keys(service)) {
+  for (const name of Object.keys(service)) {
     // tslint:disable-next-line:only-arrow-functions
-    prototype[name] = async function(request: any, options?: EnvoyGrpcRequestInit) {
-      const that = this as EnvoyClient;
-      const params = new EnvoyGrpcRequestParams(that.envoyContext, options);
-      return new Promise((resolve, reject) => {
-        constructor.prototype[name].call(
-          that,
-          request,
-          params.assembleRequestMeta(),
-          {
-            host: that.originalAddress
-          },
-          (error: ServiceError, response: any) => {
-            if (error) {
-              reject(error);
-              return;
-            }
-            resolve(response);
-          }
-        );
-      });
-    };
+    prototype[name] = makeAsyncFunc(name);
 
-    const method: any = service[name];
+    const method: any = service;
     const { originalName }: { originalName?: string } = method;
     if (originalName) {
       prototype[originalName] = prototype[name];
