+++ test/grpc.test.ts
@@ -16,13 +16,13 @@ describe("GRPC Test", () => {
       }
 
       async wrapper(call: ServerUnaryCall): Promise<any> {
-        const ctx = new EnvoyContext(call.metadata);
+        const ctx = new EnvoyContext(call.metadata, this.envoyEgressPort, GrpcTestServer.bindHost);
         expect(ctx.clientTraceId).toBe(CLIENT_TRACE_ID);
         requestId = ctx.requestId;
         traceId = ctx.traceId;
         innerParentId = ctx.spanId;
         const innerClient = new PingEnvoyClient(
-          `${GrpcTestServer.bindHost}:${this.envoyEgressPort}`,
+          `${GrpcTestServer.domainName}:${this.envoyIngressPort}`,
           ctx
         );
         return innerClient.inner({ message: call.request.message });
