+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
     "compilerOptions": {
         "moduleResolution": "node",
-        "target": "es5",
+        "target": "es6",
         "module":"es2015",
         "lib": ["es2015", "es2016", "es2017", "dom"],
         "strict": true,
-        // "sourceMap": true,
-        "inlineSourceMap": true,
+        "sourceMap": true,
+        // "inlineSourceMap": true,
         "declaration": true,
         "allowSyntheticDefaultImports": true,
         "experimentalDecorators": true,
