--- types/index.ts
@@ -1,4 +1,0 @@
-export interface HttpHeader {
-  [s: string]: string;
-}
-
